Split snub from experiment logic on unbrake-local
